+++ tests/integration/pipeline/base.py
@@ -1,13 +1,56 @@
 import os
 import re
 import shutil
+from pathlib import Path
 from typing import List
 from unittest import TestCase
 
 import boto3
 
+from tests.testing_utils import run_command_with_input
 
-class BootstrapIntegBase(TestCase):
+
+class PipelineBase(TestCase):
+    def base_command(self):
+        command = "sam"
+        if os.getenv("SAM_CLI_DEV"):
+            command = "samdev"
+
+        return command
+
+    def run_command_with_inputs(self, command_list, inputs: List[str]):
+        return run_command_with_input(command_list, ("\n".join(inputs) + "\n").encode())
+
+
+class InitIntegBase(PipelineBase):
+    generated_files: List[Path] = []
+
+    @classmethod
+    def setUpClass(cls) -> None:
+        # we need to compare the whole generated template, which is
+        # larger than normal diff size limit
+        cls.maxDiff = None
+
+    def setUp(self) -> None:
+        super().setUp()
+        self.generated_files = []
+
+    def tearDown(self) -> None:
+        for generated_file in self.generated_files:
+            if generated_file.is_dir():
+                shutil.rmtree(generated_file)
+            else:
+                generated_file.unlink()
+        super().tearDown()
+
+    def get_init_command_list(
+        self,
+    ):
+        command_list = [self.base_command(), "pipeline", "init"]
+        return command_list
+
+
+class BootstrapIntegBase(PipelineBase):
     stack_names: List[str]
 
     @classmethod
