+++ tests/integration/pipeline/test_init_command.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 
-from tests.integration.pipeline.init_integ_base import InitIntegBase
-from tests.testing_utils import run_command_with_input
+from tests.integration.pipeline.base import InitIntegBase
 
 QUICK_START_JENKINS_INPUTS = [
     "1",  # quick start
@@ -35,9 +34,7 @@ class TestInit(InitIntegBase):
         self.generated_files.append(generated_jenkinsfile_path)
 
         init_command_list = self.get_init_command_list()
-        init_process_execute = run_command_with_input(
-            init_command_list, ("\n".join(QUICK_START_JENKINS_INPUTS) + "\n").encode()
-        )
+        init_process_execute = self.run_command_with_inputs(init_command_list, QUICK_START_JENKINS_INPUTS)
 
         self.assertEqual(init_process_execute.process.returncode, 0)
         self.assertTrue(Path("Jenkinsfile").exists())
@@ -52,9 +49,7 @@ class TestInit(InitIntegBase):
         self.generated_files.append(generated_jenkinsfile_path)
 
         init_command_list = self.get_init_command_list()
-        init_process_execute = run_command_with_input(
-            init_command_list, ("\n".join(QUICK_START_JENKINS_INPUTS) + "\n").encode()
-        )
+        init_process_execute = self.run_command_with_inputs(init_command_list, QUICK_START_JENKINS_INPUTS)
 
         self.assertEqual(init_process_execute.process.returncode, 1)
         stderr = init_process_execute.stderr.decode()
@@ -68,7 +63,7 @@ class TestInit(InitIntegBase):
         inputs = ["2", str(custom_template_path), "Rainy"]  # custom template
 
         init_command_list = self.get_init_command_list()
-        init_process_execute = run_command_with_input(init_command_list, ("\n".join(inputs) + "\n").encode())
+        init_process_execute = self.run_command_with_inputs(init_command_list, inputs)
 
         self.assertEqual(init_process_execute.process.returncode, 0)
